♻️ refactor Quivr Api Key into module
