+++ backend/models/databases/repository.py
@@ -119,22 +119,6 @@ class Repository(ABC):
         pass
 
     @abstractmethod
-    def create_api_key(self, new_key_id: UUID, new_api_key: str, user_id: UUID):
-        pass
-
-    @abstractmethod
-    def delete_api_key(self, key_id: UUID, user_id: UUID):
-        pass
-
-    @abstractmethod
-    def get_active_api_key(self, api_key: UUID):
-        pass
-
-    @abstractmethod
-    def get_user_id_by_api_key(self, api_key: UUID):
-        pass
-
-    @abstractmethod
     def create_chat(self, new_chat):
         pass
