+++ backend/models/databases/supabase/__init__.py
@@ -1,5 +1,4 @@
 from models.databases.supabase.api_brain_definition import ApiBrainDefinitions
-from models.databases.supabase.api_key_handler import ApiKeyHandler
 from models.databases.supabase.brains import Brain
 from models.databases.supabase.brains_subscription_invitations import \
     BrainSubscription
