+++ backend/modules/api_key/controller/api_key_routes.py
@@ -6,24 +6,17 @@ from asyncpg.exceptions import UniqueViolationError
 from fastapi import APIRouter, Depends
 from logger import get_logger
 from middlewares.auth import AuthBearer, get_current_user
-from models import get_supabase_db
+from modules.api_key.dto.outputs import ApiKeyInfo
+from modules.api_key.entity.api_key import ApiKey
+from modules.api_key.repository.api_keys import ApiKeys
 from modules.user.entity.user_identity import UserIdentity
-from pydantic import BaseModel
 
 logger = get_logger(__name__)
 
 
-class ApiKeyInfo(BaseModel):
-    key_id: str
-    creation_time: str
-
-
-class ApiKey(BaseModel):
-    api_key: str
-    key_id: str
-
-
 api_key_router = APIRouter()
+
+api_keys_repository = ApiKeys()
 
 
 @api_key_router.post(
@@ -46,12 +39,11 @@ async def create_api_key(current_user: UserIdentity = Depends(get_current_user))
     new_key_id = uuid4()
     new_api_key = token_hex(16)
     api_key_inserted = False
-    supabase_db = get_supabase_db()
 
     while not api_key_inserted:
         try:
             # Attempt to insert new API key into database
-            supabase_db.create_api_key(new_key_id, new_api_key, current_user.id)
+            api_keys_repository.create_api_key(new_key_id, new_api_key, current_user.id)
             api_key_inserted = True
 
         except UniqueViolationError:
@@ -80,8 +72,7 @@ async def delete_api_key(
     as inactive in the database.
 
     """
-    supabase_db = get_supabase_db()
-    supabase_db.delete_api_key(key_id, current_user.id)
+    api_keys_repository.delete_api_key(key_id, current_user.id)
 
     return {"message": "API key deleted."}
 
@@ -102,6 +93,5 @@ async def get_api_keys(current_user: UserIdentity = Depends(get_current_user)):
     This endpoint retrieves all the active API keys associated with the current user. It returns a list of API key objects
     containing the key ID and creation time for each API key.
     """
-    supabase_db = get_supabase_db()
-    response = supabase_db.get_user_api_keys(current_user.id)
+    response = api_keys_repository.get_user_api_keys(current_user.id)
     return response.data
