+++ backend/modules/api_key/dto/outputs.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class ApiKeyInfo(BaseModel):
+    key_id: str
+    creation_time: str
