+++ backend/modules/api_key/entity/api_key.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class ApiKey(BaseModel):
+    api_key: str
+    key_id: str
